# BTreeSet difference, symmetric_difference & union optimized