Alternative to #65226: instead of peeking specialized to MergeIter, an additional `Peeking` intermediate. More lines of code but more readable, I think. Though it stirs a can of worms about Peekable.

It obtains the same gains for symmetric_difference and union, and also boosts `difference` a few percent (if the left set is not much smaller than the right set), e.g.
```
test difference_random_100::vs_100                     ... bench:         680 ns/iter (+/- 5)
test difference_random_100::vs_100_future              ... bench:         675 ns/iter (+/- 9)
test difference_random_10k::vs_10k                     ... bench:     166,274 ns/iter (+/- 2,515)
test difference_random_10k::vs_10k_future              ... bench:     162,993 ns/iter (+/- 1,757)
test difference_subsets::_10_vs_100                    ... bench:         213 ns/iter (+/- 1)
test difference_subsets::_10_vs_100_future             ... bench:         194 ns/iter (+/- 5)
```
In addition, some cleanup of earlier changes suggested by clippy, which I can push to #65226 too.

r? @bluss