@bluss 2019-10-13
Nice benchmarks (posted in the other pr)! Looks like a neat improvement