@ssomers 2019-10-14
These are the readings for the updated PR:
```
test difference_random_100::vs_100                     ... bench:         729 ns/iter (+/- 14)
test difference_random_100::vs_100_future              ... bench:         686 ns/iter (+/- 28)
test difference_random_10k::vs_10k                     ... bench:     168,942 ns/iter (+/- 5,895)
test difference_random_10k::vs_10k_future              ... bench:     160,226 ns/iter (+/- 4,405)
test difference_subsets::_10_vs_100                    ... bench:         214 ns/iter (+/- 9)
test difference_subsets::_10_vs_100_future             ... bench:         187 ns/iter (+/- 4)
test symmdiff_neg_vs_pos::_100_neg_vs_100_pos          ... bench:         880 ns/iter (+/- 123)
test symmdiff_neg_vs_pos::_100_neg_vs_100_pos_future   ... bench:         890 ns/iter (+/- 32)
test symmdiff_neg_vs_pos::_100_neg_vs_10k_pos          ... bench:      37,187 ns/iter (+/- 980)
test symmdiff_neg_vs_pos::_100_neg_vs_10k_pos_future   ... bench:      36,721 ns/iter (+/- 1,002)
test symmdiff_neg_vs_pos::_100_pos_vs_100_neg          ... bench:         901 ns/iter (+/- 26)
test symmdiff_neg_vs_pos::_100_pos_vs_100_neg_future   ... bench:         856 ns/iter (+/- 35)
test symmdiff_neg_vs_pos::_100_pos_vs_10k_neg          ... bench:      44,593 ns/iter (+/- 965)
test symmdiff_neg_vs_pos::_100_pos_vs_10k_neg_future   ... bench:      43,231 ns/iter (+/- 1,275)
test symmdiff_neg_vs_pos::_10k_neg_vs_100_pos          ... bench:      49,806 ns/iter (+/- 2,459)
test symmdiff_neg_vs_pos::_10k_neg_vs_100_pos_future   ... bench:      45,261 ns/iter (+/- 3,033)
test symmdiff_neg_vs_pos::_10k_neg_vs_10k_pos          ... bench:      88,181 ns/iter (+/- 4,752)
test symmdiff_neg_vs_pos::_10k_neg_vs_10k_pos_future   ... bench:      84,279 ns/iter (+/- 3,806)
test symmdiff_neg_vs_pos::_10k_pos_vs_100_neg          ... bench:      43,414 ns/iter (+/- 2,347)
test symmdiff_neg_vs_pos::_10k_pos_vs_100_neg_future   ... bench:      39,981 ns/iter (+/- 2,312)
test symmdiff_neg_vs_pos::_10k_pos_vs_10k_neg          ... bench:      87,481 ns/iter (+/- 3,478)
test symmdiff_neg_vs_pos::_10k_pos_vs_10k_neg_future   ... bench:      82,290 ns/iter (+/- 2,926)
test symmdiff_random_100::vs_100                       ... bench:         972 ns/iter (+/- 37)
test symmdiff_random_100::vs_100_future                ... bench:         953 ns/iter (+/- 41)
test symmdiff_random_100::vs_10k                       ... bench:      56,538 ns/iter (+/- 2,140)
test symmdiff_random_100::vs_10k_future                ... bench:      59,465 ns/iter (+/- 1,742)
test symmdiff_random_100::vs_1600                      ... bench:       9,970 ns/iter (+/- 264)
test symmdiff_random_100::vs_1600_future               ... bench:       9,692 ns/iter (+/- 502)
test symmdiff_random_10k::vs_10k                       ... bench:     209,957 ns/iter (+/- 13,843)
test symmdiff_random_10k::vs_10k_future                ... bench:     205,140 ns/iter (+/- 9,867)
test symmdiff_subsets::_100_vs_10k                     ... bench:      41,016 ns/iter (+/- 1,358)
test symmdiff_subsets::_100_vs_10k_future              ... bench:      39,083 ns/iter (+/- 1,158)
test symmdiff_subsets::_10_vs_100                      ... bench:         427 ns/iter (+/- 25)
test symmdiff_subsets::_10_vs_100_future               ... bench:         433 ns/iter (+/- 22)
test union_neg_vs_pos::_100_neg_vs_100_pos             ... bench:         791 ns/iter (+/- 41)
test union_neg_vs_pos::_100_neg_vs_100_pos_future      ... bench:         902 ns/iter (+/- 14)
test union_neg_vs_pos::_100_neg_vs_10k_pos             ... bench:      36,296 ns/iter (+/- 3,422)
test union_neg_vs_pos::_100_neg_vs_10k_pos_future      ... bench:      42,735 ns/iter (+/- 2,068)
test union_neg_vs_pos::_100_pos_vs_100_neg             ... bench:         849 ns/iter (+/- 34)
test union_neg_vs_pos::_100_pos_vs_100_neg_future      ... bench:         915 ns/iter (+/- 43)
test union_neg_vs_pos::_100_pos_vs_10k_neg             ... bench:      41,275 ns/iter (+/- 5,920)
test union_neg_vs_pos::_100_pos_vs_10k_neg_future      ... bench:      42,866 ns/iter (+/- 1,007)
test union_neg_vs_pos::_10k_neg_vs_100_pos             ... bench:      43,137 ns/iter (+/- 1,527)
test union_neg_vs_pos::_10k_neg_vs_100_pos_future      ... bench:      40,187 ns/iter (+/- 1,012)
test union_neg_vs_pos::_10k_neg_vs_10k_pos             ... bench:      78,250 ns/iter (+/- 2,655)
test union_neg_vs_pos::_10k_neg_vs_10k_pos_future      ... bench:      83,164 ns/iter (+/- 2,745)
test union_neg_vs_pos::_10k_pos_vs_100_neg             ... bench:      44,690 ns/iter (+/- 1,496)
test union_neg_vs_pos::_10k_pos_vs_100_neg_future      ... bench:      43,874 ns/iter (+/- 1,177)
test union_neg_vs_pos::_10k_pos_vs_10k_neg             ... bench:      85,101 ns/iter (+/- 1,054)
test union_neg_vs_pos::_10k_pos_vs_10k_neg_future      ... bench:      85,770 ns/iter (+/- 2,906)
test union_random_100::vs_100                          ... bench:         942 ns/iter (+/- 39)
test union_random_100::vs_100_future                   ... bench:         936 ns/iter (+/- 45)
test union_random_100::vs_10k                          ... bench:      49,211 ns/iter (+/- 1,340)
test union_random_100::vs_10k_future                   ... bench:      50,612 ns/iter (+/- 1,617)
test union_random_100::vs_1600                         ... bench:       8,754 ns/iter (+/- 312)
test union_random_100::vs_1600_future                  ... bench:       8,708 ns/iter (+/- 227)
test union_random_10k::vs_10k                          ... bench:     199,802 ns/iter (+/- 5,530)
test union_random_10k::vs_10k_future                   ... bench:     191,034 ns/iter (+/- 3,066)
test union_subsets::_100_vs_10k                        ... bench:      36,560 ns/iter (+/- 941)
test union_subsets::_100_vs_10k_future                 ... bench:      40,408 ns/iter (+/- 1,325)
test union_subsets::_10_vs_100                         ... bench:         422 ns/iter (+/- 8)
test union_subsets::_10_vs_100_future                  ... bench:         449 ns/iter (+/- 23)
```
Theoretically (and in practice) unchanged tests pruned from this list. You see that towards the end, the future gets bleak. While the future is bright when I build & run only the MergeIter-based benches  (using --features merge), as I've always been doing lately. Each run is quite reproducible: the more benches built, the slower the they all become, but those named union_*_future suffer much more. Running fewer benches with --benches gives the same timing as running all, so it's the build that is different. Seems like the more code there is mixed in, the less the optimizer focuses, and the future code needs more attention.

Anyways, looking into cargo-benchcmp now.